+++ websockets/test_client_server.py
@@ -8,7 +8,7 @@ import unittest
 import unittest.mock
 
 from .client import *
-from .exceptions import ConnectionClosed, InvalidHandshake
+from .exceptions import ConnectionClosed, InvalidHandshake, InvalidStatus
 from .http import USER_AGENT, read_response
 from .server import *
 
@@ -251,7 +251,7 @@ class ClientServerTests(unittest.TestCase):
 
     def test_authentication(self):
         self.start_server(klass=ForbiddenWebSocketServerProtocol)
-        with self.assertRaises(InvalidHandshake):
+        with self.assertRaises(InvalidStatus):
             self.start_client()
         self.stop_server()
 
@@ -360,7 +360,7 @@ class ClientServerTests(unittest.TestCase):
         _read_response.side_effect = wrong_read_response
 
         self.start_server()
-        with self.assertRaises(InvalidHandshake):
+        with self.assertRaises(InvalidStatus):
             self.start_client()
         self.run_loop_once()
         self.stop_server()
@@ -418,7 +418,7 @@ class ClientServerTests(unittest.TestCase):
         self.stop_server()
 
         # Opening handshake fails with 503 Service Unavailable
-        self.assertEqual(str(raised.exception), "Bad status code: 503")
+        self.assertEqual(str(raised.exception), "Status code not 101: 503")
 
     def test_server_shuts_down_during_connection_handling(self):
         self.start_server()
@@ -432,6 +432,15 @@ class ClientServerTests(unittest.TestCase):
 
         # Websocket connection terminates with 1001 Going Away.
         self.assertEqual(self.client.close_code, 1001)
+
+    def test_invalid_status_error_during_client_connect(self):
+        self.start_server(klass=ForbiddenWebSocketServerProtocol)
+        with self.assertRaises(InvalidStatus) as raised:
+            self.start_client()
+        exception = raised.exception
+        self.assertEqual(str(exception), "Status code not 101: 403")
+        self.assertEqual(exception.code, 403)
+        self.stop_server()
 
     @unittest.mock.patch('websockets.server.read_request')
     def test_connection_error_during_opening_handshake(self, _read_request):
@@ -522,7 +531,8 @@ class ClientServerOriginTests(unittest.TestCase):
     def test_checking_origin_fails(self):
         server = self.loop.run_until_complete(
             serve(handler, 'localhost', 8642, origins=['http://localhost']))
-        with self.assertRaisesRegex(InvalidHandshake, "Bad status code: 403"):
+        with self.assertRaisesRegex(InvalidHandshake,
+                                    "Status code not 101: 403"):
             self.loop.run_until_complete(
                 connect('ws://localhost:8642/', origin='http://otherhost'))