+++ docs/changelog.rst
@@ -16,6 +16,10 @@ Changelog
 
 * Rewrote HTTP handling for simplicity and performance.
 
+* An invalid response status code during :func:`~websockets.client.connect`
+  now raises :class:`~websockets.exceptions.InvalidStatus` with a ``code``
+  attribute.
+
 3.3
 ...